NAS-114991: Moved ix-drop-grid
