+++ src/app/modules/ix-drop-grid/ix-drag.directive.ts
@@ -18,13 +18,13 @@ import {
 } from '@angular/core';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { take } from 'rxjs/operators';
-import { IxDragHandleDirective } from 'app/pages/common/ix-drop-grid/ix-drag-handle.directive';
-import { IxDropGridItemDirective } from 'app/pages/common/ix-drop-grid/ix-drop-grid-item.directive';
+import { IxDragHandleDirective } from 'app/modules/ix-drop-grid/ix-drag-handle.directive';
+import { IxDropGridItemDirective } from 'app/modules/ix-drop-grid/ix-drop-grid-item.directive';
 import {
   ixDragHandleDirectiveToken,
   ixDragParentToken,
   ixDropGridItemDirectiveToken,
-} from 'app/pages/common/ix-drop-grid/ix-drop-grid.tokens';
+} from 'app/modules/ix-drop-grid/ix-drop-grid.tokens';
 
 @UntilDestroy()
 @Directive({
